Update vimrc and bashrc
bashrc:

I added a way to tell which shell level ($SHLVL) you are working in.  I
also added an alias called "anchor" which adds the number of the
top-level shell ($TOPLVL) and an anchor character to the prompt.

vimrc:

Tweaked netrw behavior to get rid of [RO] buffers which sometimes make
exiting difficult.